restore: remove mons command and move to restore
to make it easier to find commands, `mons restore-quorum`
command is moved to `restore mon-quorum`.
