+++ cmd/odf/restore/mon_quorum.go
@@ -1,4 +1,4 @@
-package mon
+package restore
 
 import (
 	"github.com/red-hat-storage/odf-cli/cmd/odf/root"
@@ -6,13 +6,13 @@ import (
 	"github.com/spf13/cobra"
 )
 
-// RestoreQuorum represents the mons command
-var RestoreQuorum = &cobra.Command{
-	Use:                "restore-quorum",
+// monQuorumCmd represents the mons command
+var monQuorumCmd = &cobra.Command{
+	Use:                "mon-quorum",
 	Short:              "When quorum is lost, restore quorum to the remaining healthy mon",
 	DisableFlagParsing: true,
 	Args:               cobra.ExactArgs(1),
-	Example:            "odf mons restore-quorum <mon_id>",
+	Example:            "odf restore mon-quorum <mon_id>",
 	Run: func(cmd *cobra.Command, args []string) {
 		mons.RestoreQuorum(cmd.Context(), root.ClientSets, root.OperatorNamespace, root.StorageClusterNamespace, args[0])
 	},
